refactor: Move the INode schemas to n8n-workflow
